7 again

+++ ex_7.py
@@ -1,4 +1,4 @@
-n1 = int(input('first note:\n'))
-n2 = int(input('secound note:\n'))
+n1 = float(input('first note:\n'))
+n2 = float(input('secound note:\n'))
 m = (n1 + n2) /2
 print('a média é {}!'.format(m))
